requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose
passlib[bcrypt]
python-dotenv
pytest
slowapi
psycopg2-binary

